requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
imbalanced-learn>=0.8.0

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0
timm>=0.4.12

# API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
shap>=0.40.0
opencv-python>=4.5.0
Pillow>=8.3.0

# Utilities
tqdm>=4.62.0
joblib>=1.0.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0
